refac: reorganized code structure
This is a huge commit / change. It's a complete rewrite. The code is now much more clearly organized into seperate modules. The `main.py` now just holds the pipeline-functions to execute the validation (and not everything like before). Furthermore a Output type (class) was introduced, which holds structured information on the schematron output.